% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{aggiungi_colonne_base_dt}
\alias{aggiungi_colonne_base_dt}
\title{Aggiunge Colonne Log2 e Flag High/Low Base}
\usage{
aggiungi_colonne_base_dt(counts_dt)
}
\arguments{
\item{counts_dt}{Una data.table con la colonna 'count'. Viene modificata in-place.}
}
\value{
La data.table modificata con le colonne 'log2_counts' e 'high' aggiunte.
}
\description{
Aggiunge in-place due nuove colonne a una data.table di conteggi:
'log2_counts' (log2(count + 1)) e un flag binario 'high' (1 se count > 100).
}
