% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{aggrega_conteggi_per_paziente_dt}
\alias{aggrega_conteggi_per_paziente_dt}
\title{Aggrega i Conteggi Totali per Paziente}
\usage{
aggrega_conteggi_per_paziente_dt(counts_dt, meta_dt)
}
\arguments{
\item{counts_dt}{Una data.table con 'sample_id' e 'count'.}

\item{meta_dt}{Una data.table con 'sample_id' e 'patient_id'.}
}
\value{
Una data.table con 'patient_id' e 'total_count'.
}
\description{
Esegue il join dei conteggi con i metadati e calcola la somma totale
dei conteggi per ciascun `patient_id`.
}
