% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{bench_indice_dt}
\alias{bench_indice_dt}
\title{Benchmark di Query con e senza Indice Secondario}
\usage{
bench_indice_dt(counts_dt, target_gene, target_sample)
}
\arguments{
\item{counts_dt}{Una data.table con 'gene' e 'sample_id'.}

\item{target_gene}{Il nome del gene da cercare.}

\item{target_sample}{L'ID del campione da cercare.}
}
\value{
Una lista contenente i tempi di esecuzione e il risultato della query con indice.
}
\description{
Confronta il tempo di esecuzione di una query complessa prima e dopo
la creazione di un Indice Secondario (`setindex()`).
NOTA: Richiede il caricamento del pacchetto `tictoc` per funzionare correttamente.
}
