% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{calcola_correlazione_crp_vitals_dt}
\alias{calcola_correlazione_crp_vitals_dt}
\title{Calcola la Correlazione tra CRP e Vital Signs}
\usage{
calcola_correlazione_crp_vitals_dt(matched_dt)
}
\arguments{
\item{matched_dt}{Una data.table risultante da `nearest_time_matching_dt`.}
}
\value{
Una data.table con 'patient_id', 'corr_CRP_HR' e 'corr_CRP_SBP'.
}
\description{
Filtra il set di dati abbinato per 'CRP' e calcola la correlazione di Pearson
tra il valore di laboratorio (CRP) e i vital signs (HR e SBP) per ciascun paziente.
}
