% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{calcola_media_gene_condition_dt}
\alias{calcola_media_gene_condition_dt}
\title{Calcola la Media dei Conteggi per Gene e Condition}
\usage{
calcola_media_gene_condition_dt(gene_counts_dt, sample_dt)
}
\arguments{
\item{gene_counts_dt}{Una data.table con colonne 'gene', 'sample_id', 'count'.}

\item{sample_dt}{Una data.table con colonne 'sample_id', 'condition'.}
}
\value{
Una data.table con 'gene', 'condition' e 'mean_count'.
}
\description{
Unisce i conteggi con i metadati e calcola la media dei conteggi
raggruppando per gene e condizione (treatment/control).
}
