% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{calcola_media_mediana_geni_00_dt}
\alias{calcola_media_mediana_geni_00_dt}
\title{Calcola Media e Mediana dei Conteggi per Geni 'GENE_00' Trattati}
\usage{
calcola_media_mediana_geni_00_dt(gene_counts_dt, sample_dt)
}
\arguments{
\item{gene_counts_dt}{Una data.table con colonne 'gene', 'sample_id', 'count'.}

\item{sample_dt}{Una data.table con colonne 'sample_id', 'condition'.}
}
\value{
Una data.table riepilogativa con 'gene', 'mean_count' e 'median_count'.
}
\description{
Filtra i campioni 'treated' e calcola la media e la mediana dei conteggi
per i geni che iniziano con 'GENE_00', raggruppando per nome del gene.
}
