% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{calcola_media_per_gene_dt}
\alias{calcola_media_per_gene_dt}
\title{Calcola la Media dei Conteggi per Gene}
\usage{
calcola_media_per_gene_dt(dt_long)
}
\arguments{
\item{dt_long}{Una data.table in formato long con le colonne 'gene' e 'Count'.}
}
\value{
Una data.table con 'gene' e 'Mean_Count', ordinata per gene.
}
\description{
Calcola la media dei conteggi (`Mean_Count`) per ciascun gene in un data.table
in formato long, ordinando il risultato per nome del gene.
}
