% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{calcola_media_top_varianti_dt}
\alias{calcola_media_top_varianti_dt}
\title{Calcola la Media dei Conteggi per Gruppo per i Top Geni Variabili}
\usage{
calcola_media_top_varianti_dt(combined_samples, bulk_counts_long)
}
\arguments{
\item{combined_samples}{Data.table combinata dei metadati ('sample_id', 'cohort', 'condition').}

\item{bulk_counts_long}{Data.table di conteggi in formato long ('sample_id', 'gene', 'count').}
}
\value{
Una data.table con le medie dei conteggi per i Top 100 geni variabili per ogni gruppo.
}
\description{
Calcola la varianza per gene, seleziona i Top 100 più variabili, filtra i conteggi,
unisce i dati e calcola la media dei conteggi raggruppando per coorte, condizione e gene.
}
