% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{calcola_somma_geni_00_dt}
\alias{calcola_somma_geni_00_dt}
\title{Calcola la Somma Totale dei Conteggi per 'GENE_00' Trattati}
\usage{
calcola_somma_geni_00_dt(gene_counts_dt, sample_dt)
}
\arguments{
\item{gene_counts_dt}{Una data.table con colonne 'gene', 'sample_id', 'count'.}

\item{sample_dt}{Una data.table con colonne 'sample_id', 'condition'.}
}
\value{
Una data.table riepilogativa con 'sample_id' e 'total_count_GENE_00'.
}
\description{
Filtra i campioni 'treated' e calcola la somma totale dei conteggi per tutti
i geni il cui nome inizia con 'GENE_00', raggruppando per ID campione.
}
