% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{calcola_sovrapposizione_totale_dt}
\alias{calcola_sovrapposizione_totale_dt}
\title{Calcola la Somma Totale di Basi Sovrapposte per Gene}
\usage{
calcola_sovrapposizione_totale_dt(overlap_dt)
}
\arguments{
\item{overlap_dt}{Una data.table risultante da `interseca_geni_e_picchi_dt`.}
}
\value{
Una data.table con 'gene' e 'total_overlap_bp'.
}
\description{
Calcola l'intervallo di sovrapposizione effettivo (in bp) per ciascuna coppia
picco-gene e poi aggrega la somma totale di bp sovrapposte per ciascun gene.
}
