% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{calcola_statistiche_wide_dt}
\alias{calcola_statistiche_wide_dt}
\title{Calcola Statistiche Riassuntive e Trasforma in Wide}
\usage{
calcola_statistiche_wide_dt(counts_dt, metadata_dt)
}
\arguments{
\item{counts_dt}{Una data.table con 'sample_id', 'gene', 'count'.}

\item{metadata_dt}{Una data.table con 'sample_id', 'condition'.}
}
\value{
Una data.table in formato wide con le statistiche per ogni condizione.
}
\description{
Calcola media, mediana, Q1, Q3 per 'gene' e 'condition', quindi trasforma
la data.table dal formato long a wide (`dcast`).
}
