% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{calcola_tassi_anomali_dt}
\alias{calcola_tassi_anomali_dt}
\title{Calcola i Tassi di Risultati Anomali per Paziente e Lab}
\usage{
calcola_tassi_anomali_dt(classified_labs)
}
\arguments{
\item{classified_labs}{Una data.table con 'patient_id', 'lab', 'status'.}
}
\value{
Una data.table riepilogativa ordinata per 'abnormal_rate' decrescente.
}
\description{
Aggrega i risultati classificati, calcolando il conteggio totale, il conteggio anomalo
e il tasso percentuale di risultati anomali per ciascuna coppia paziente-lab.
}
