% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{carica_e_prepara_dati}
\alias{carica_e_prepara_dati}
\title{Carica, Pulisce e Unisce i Dati di Integrazione e Annotazione}
\usage{
carica_e_prepara_dati(integration_file, annotation_file)
}
\arguments{
\item{integration_file}{Percorso del file CSV dei dati di integrazione.}

\item{annotation_file}{Percorso del file CSV dei dati di annotazione.}
}
\value{
Una data.table combinata, pulita e riordinata.
}
\description{
Carica i dati di integrazione e annotazione, pulisce la colonna 'cell'
(rimuovendo pattern come '_XY_') e unisce le due data.table.
}
