% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{conta_picchi_per_gene_dt}
\alias{conta_picchi_per_gene_dt}
\title{Conta i Picchi Unici Sovrapposti per Gene}
\usage{
conta_picchi_per_gene_dt(overlap_dt)
}
\arguments{
\item{overlap_dt}{Una data.table risultante da `interseca_geni_e_picchi_dt`.}
}
\value{
Una data.table con 'gene' e 'num_peaks'.
}
\description{
Calcola il numero di picchi ATAC unici che si sovrappongono a ciascun gene,
utilizzando la funzione `uniqueN()`.
}
