% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{esegui_equi_join_dt}
\alias{esegui_equi_join_dt}
\title{Esegue un Equi-Join con setkey()}
\usage{
esegui_equi_join_dt(counts_dt, metadata_dt)
}
\arguments{
\item{counts_dt}{Una data.table con i conteggi (colonna 'sample_id').}

\item{metadata_dt}{Una data.table con i metadati del campione (colonna 'sample_id').}
}
\value{
Una data.table unita (joined_dt).
}
\description{
Esegue un inner join tra i conteggi dei geni e i metadati del campione (su 'sample_id'),
sfruttando `setkey()` sul data.table dei metadati per ottimizzare il join.
}
