% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{filtra_geni_differenziali_dt}
\alias{filtra_geni_differenziali_dt}
\title{Filtra Geni con Espressione Differenziale}
\usage{
filtra_geni_differenziali_dt(wide_stats_dt)
}
\arguments{
\item{wide_stats_dt}{Una data.table in formato wide con colonne statistiche per 'treated' e 'control'.}
}
\value{
Una data.table con i geni differenziali filtrati e le relative medie.
}
\description{
Filtra una data.table in formato wide (risultato di `calcola_statistiche_wide_dt`)
per identificare i geni in cui la media trattata (`mean_treated`) è almeno il doppio
della media di controllo (`mean_control`).
}
