% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{filtra_picchi_atac_dt}
\alias{filtra_picchi_atac_dt}
\title{Filtra i Picchi ATAC per Posizione Genomica}
\usage{
filtra_picchi_atac_dt(atac_peaks_dt)
}
\arguments{
\item{atac_peaks_dt}{Una data.table con 'chr', 'start', 'score'.}
}
\value{
Una data.table contenente solo i picchi filtrati.
}
\description{
Filtra una data.table di picchi ATAC per un cromosoma specifico ('chr2') e
un intervallo genomico definito (2M-4M bp).
}
