% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{geni_in_tutti_i_campioni_dt}
\alias{geni_in_tutti_i_campioni_dt}
\title{Identifica Geni con Varianti HIGH-Impact in Tutti i Campioni}
\usage{
geni_in_tutti_i_campioni_dt(variants_genes_dt)
}
\arguments{
\item{variants_genes_dt}{Una data.table di varianti mappate, inclusa la colonna 'impact'.}
}
\value{
Una lista contenente `result` (data.table dei geni) e `total_samples` (conteggio totale).
}
\description{
Trova i geni che hanno almeno una variante HIGH-Impact presente in tutti i campioni
unici nel set di dati.
}
