% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{interseca_geni_e_picchi_dt}
\alias{interseca_geni_e_picchi_dt}
\title{Intersezione Genomica tra Picchi e Geni (foverlaps)}
\usage{
interseca_geni_e_picchi_dt(atac_dt, gene_dt)
}
\arguments{
\item{atac_dt}{Una data.table con intervalli di picchi ('chr', 'start', 'end', 'peak_id').}

\item{gene_dt}{Una data.table con intervalli di geni ('chr', 'start', 'end', 'gene').}
}
\value{
Una data.table contenente tutte le coppie picco-gene sovrapposte.
}
\description{
Esegue un'intersezione genomica utilizzando `foverlaps()` per trovare le sovrapposizioni
tra intervalli di picchi ATAC e intervalli di geni.
}
