% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{mappa_varianti_a_geni_dt}
\alias{mappa_varianti_a_geni_dt}
\title{Mappa Varianti Genomiche a Intervalli Genici}
\usage{
mappa_varianti_a_geni_dt(variants_dt, genes_dt)
}
\arguments{
\item{variants_dt}{Una data.table di varianti ('chr', 'pos', 'ref', 'alt', 'impact', 'sample_id').}

\item{genes_dt}{Una data.table di geni ('chr', 'start', 'end', 'gene').}
}
\value{
Una data.table di varianti mappate ai geni.
}
\description{
Mappa le posizioni delle varianti (considerate come intervalli 1-bp) agli
intervalli dei geni, utilizzando `foverlaps()` con `type = "within"` (la variante
deve essere contenuta nel gene).
}
