% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{ordina_campioni_dt}
\alias{ordina_campioni_dt}
\title{Ordina Campioni per Coorte, Condizione e ID}
\usage{
ordina_campioni_dt(combined_samples)
}
\arguments{
\item{combined_samples}{La data.table risultante da `unisci_metadati_dt`.}
}
\value{
La data.table ordinata (modificata in-place).
}
\description{
Ordina in-place la data.table combinata per le colonne 'cohort', 'condition'
e 'sample_id'.
}
