% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{ricalcola_high_per_gene_dt}
\alias{ricalcola_high_per_gene_dt}
\title{Ricalcola il Flag 'High' Basato sulla Mediana del Gene}
\usage{
ricalcola_high_per_gene_dt(counts_dt_with_cols)
}
\arguments{
\item{counts_dt_with_cols}{Una data.table con le colonne 'gene' e 'count'. Viene modificata in-place.}
}
\value{
La data.table modificata con il flag 'high' ricalcolato per gene.
}
\description{
Ricalcola in-place il flag 'high' a 1 se il conteggio è superiore alla
mediana del conteggio per quel gene specifico.
}
