% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{seleziona_top_picchi_dt}
\alias{seleziona_top_picchi_dt}
\title{Seleziona i Top 50 Picchi per Score}
\usage{
seleziona_top_picchi_dt(filtered_peaks)
}
\arguments{
\item{filtered_peaks}{Una data.table di picchi ATAC (risultato di `filtra_picchi_atac_dt`).}
}
\value{
Una data.table con le prime 50 righe ordinate per score.
}
\description{
Ordina in-place i picchi filtrati per 'score' decrescente e seleziona i Top 50.
}
