% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{task1_salva_dati_combinati}
\alias{task1_salva_dati_combinati}
\title{Salva il Data.table Combinato su File CSV}
\usage{
task1_salva_dati_combinati(
  combined_dt,
  output_file = "combined_analysis_data.csv"
)
}
\arguments{
\item{combined_dt}{La data.table combinata da salvare.}

\item{output_file}{Il percorso e nome del file di output (default: "combined_analysis_data.csv").}
}
\value{
Il percorso del file di output.
}
\description{
Salva il data.table risultante dalla pre-elaborazione su un file CSV specificato.
}
