% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{task2_conta_cellule_per_cluster_celltype}
\alias{task2_conta_cellule_per_cluster_celltype}
\title{Calcola il Conteggio Cellule per Cluster e Cell Type}
\usage{
task2_conta_cellule_per_cluster_celltype(
  combined_dt,
  output_file = "cell_type_counts_per_cluster.csv"
)
}
\arguments{
\item{combined_dt}{La data.table combinata con le colonne di raggruppamento.}

\item{output_file}{Il percorso e nome del file di output (default: "cell_type_counts_per_cluster.csv").}
}
\value{
La data.table dei conteggi.
}
\description{
Calcola il conteggio (`.N`) delle cellule raggruppato per `integration_cluster` e `cell_type`,
e salva il risultato su un file CSV.
}
