% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{task3_riepilogo_cluster_celltype_tissue}
\alias{task3_riepilogo_cluster_celltype_tissue}
\title{Calcola il Riepilogo Cellule per Cluster, Cell Type e Tissue}
\usage{
task3_riepilogo_cluster_celltype_tissue(
  combined_dt,
  output_file = "summary_cluster_celltype_tissue.csv"
)
}
\arguments{
\item{combined_dt}{La data.table combinata con le colonne di raggruppamento.}

\item{output_file}{Il percorso e nome del file di output (default: "summary_cluster_celltype_tissue.csv").}
}
\value{
La data.table del riepilogo.
}
\description{
Calcola il conteggio (`.N`) delle cellule raggruppato per `integration_cluster`,
`cell_type` e `sample_type`, e salva il risultato su un file CSV.
}
