% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{task4_crea_grafico_distribuzione}
\alias{task4_crea_grafico_distribuzione}
\title{Crea un Grafico a Barre della Distribuzione del Tessuto}
\usage{
task4_crea_grafico_distribuzione(
  summary_table,
  output_png = "cell_type_distribution_by_cluster_and_tissue.png"
)
}
\arguments{
\item{summary_table}{La data.table riepilogativa (risultato di `task3_riepilogo_cluster_celltype_tissue`).}

\item{output_png}{Il nome del file PNG in cui salvare il grafico (default: "cell_type_distribution_by_cluster_and_tissue.png").}
}
\value{
L'oggetto grafico `ggplot`.
}
\description{
Calcola la proporzione del tipo di tessuto (N/T) all'interno di ciascun gruppo
(cell type x cluster) e genera un grafico a barre impilate utilizzando `ggplot2`.
}
