% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{trasforma_e_calcola_totale_dt}
\alias{trasforma_e_calcola_totale_dt}
\title{Trasforma Dati Wide in Long e Calcola il Totale Campione}
\usage{
trasforma_e_calcola_totale_dt(bulk_wide)
}
\arguments{
\item{bulk_wide}{Una data.table con la colonna 'gene' e colonne per ciascun campione.}
}
\value{
Una data.table in formato long con colonne 'gene', 'Sample', 'Count', 'Sample_Total'.
}
\description{
Trasforma una tabella di conteggi da wide a long (`melt`) e calcola il
conteggio totale di espressione (`Sample_Total`) per ciascun campione.
}
