% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{trova_top_geni_overlap_dt}
\alias{trova_top_geni_overlap_dt}
\title{Trova i Top 20 Geni per Lunghezza Totale di Sovrapposizione}
\usage{
trova_top_geni_overlap_dt(total_overlap_dt, peaks_per_gene_dt)
}
\arguments{
\item{total_overlap_dt}{Risultato di `calcola_sovrapposizione_totale_dt`.}

\item{peaks_per_gene_dt}{Risultato di `conta_picchi_per_gene_dt`.}
}
\value{
Una data.table con i Top 20 geni e i loro valori di sovrapposizione.
}
\description{
Unisce i riepiloghi dei conteggi dei picchi e della lunghezza totale di
sovrapposizione, quindi seleziona i 20 geni con la sovrapposizione totale più alta.
}
