% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{trova_top_geni_per_condition_dt}
\alias{trova_top_geni_per_condition_dt}
\title{Trova i Top 10 Geni per Condition in base alla Media}
\usage{
trova_top_geni_per_condition_dt(counts_dt, meta_dt)
}
\arguments{
\item{counts_dt}{Una data.table con 'sample_id', 'gene', 'count'.}

\item{meta_dt}{Una data.table con 'sample_id', 'condition'.}
}
\value{
Una data.table con i Top 10 geni per ogni condizione ('condition', 'gene', 'avg_count').
}
\description{
Calcola la media dei conteggi per gene e condition, quindi seleziona i
Top 10 geni con la media più alta all'interno di ciascuna condizione.
}
