% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONScomm.R
\name{unisci_metadati_dt}
\alias{unisci_metadati_dt}
\title{Unisce Metadati di Coorti Multiple}
\usage{
unisci_metadati_dt(cohortA_dt, cohortB_dt)
}
\arguments{
\item{cohortA_dt}{Data.table dei metadati della coorte A.}

\item{cohortB_dt}{Data.table dei metadati della coorte B.}
}
\value{
Una data.table combinata con i metadati di entrambe le coorti.
}
\description{
Unisce le data.table di metadati da coorti diverse utilizzando `rbindlist`,
gestendo eventuali differenze di colonne (`fill = TRUE`).
}
